[*] Переделана архитектура коллекции [f] Исправлено представление поиска

+++ RJTranslate-App/Models/UIColor+RJT_Private.m
@@ -14,9 +14,9 @@
 {
     CGFloat r, g, b, a;
     if ([self getRed:&r green:&g blue:&b alpha:&a])
-        return [UIColor colorWithRed:MAX(r - 0.2, 0.0)
-                               green:MAX(g - 0.2, 0.0)
-                                blue:MAX(b - 0.2, 0.0)
+        return [UIColor colorWithRed:(CGFloat)MAX(r - 0.2, 0.0)
+                               green:(CGFloat)MAX(g - 0.2, 0.0)
+                                blue:(CGFloat)MAX(b - 0.2, 0.0)
                                alpha:a];
     return self;
 }
